+++ 20230702-poi-practice/server/src/main/kotlin/com/usktea/demo/services/PaymentService.kt
@@ -21,12 +21,15 @@ class PaymentService(
 
     @Transactional
     fun addPayments(): Long {
-        for (i: Int in 1..500000) {
+        val currentTotal = paymentRepository.count()
+
+        for (i: Int in 1..10000) {
+            val current = currentTotal + i
             paymentRepository.save(
                 Payment(
                     price = i * 1000.toLong(),
-                    amount = i.toLong(),
-                    productName = "상품$i"
+                    amount = current,
+                    productName = "상품$current"
                 )
             )
         }
@@ -44,54 +47,54 @@ class PaymentService(
         return paymentRepository.findAll(page)
     }
 
-    fun paymentsToExcel(payments: MutableList<Payment>): File? {
-
-        try {
-            val createDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"))
-            val workbook: Workbook = XSSFWorkbook()
-
-            val sheet = workbook.createSheet("Payment_$createDate")
-
-            val headerFont = workbook.createFont()
-            headerFont.bold = true
-            headerFont.fontHeightInPoints = 14
-
-            val headerCellStyle = workbook.createCellStyle()
-            headerCellStyle.setFont(headerFont)
-            val headerRow = sheet.createRow(0)
-            val columns = arrayOf(
-                "Price",
-                "amount",
-                "ProductName"
-            )
-            for (i in columns.indices) {
-                val cell = headerRow.createCell(i)
-                cell.setCellValue(columns[i])
-                cell.cellStyle = headerCellStyle
-            }
-
-            var rowNum = 1
-            for (payment in payments) {
-                val row = sheet.createRow(rowNum)
-                row.createCell(0).setCellValue(payment.price.toDouble())
-                row.createCell(1).setCellValue(payment.amount.toDouble())
-                row.createCell(2).setCellValue(payment.productName)
-                rowNum++
-            }
-
-            for (i in columns.indices) {
-                sheet.autoSizeColumn(i)
-            }
-
-            val file = File.createTempFile("payments_${createDate}", ".xlsx")
-            file.deleteOnExit()
-            val fileOut = FileOutputStream(file)
-            workbook.write(fileOut)
-            fileOut.close()
-            workbook.close()
-            return file
-        } catch (e: IOException) {
-            throw IOException(e)
-        }
-    }
+//    fun paymentsToExcel(payments: MutableList<Payment>): File? {
+//
+//        try {
+//            val createDate = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd"))
+//            val workbook: Workbook = XSSFWorkbook()
+//
+//            val sheet = workbook.createSheet("Payment_$createDate")
+//
+//            val headerFont = workbook.createFont()
+//            headerFont.bold = true
+//            headerFont.fontHeightInPoints = 14
+//
+//            val headerCellStyle = workbook.createCellStyle()
+//            headerCellStyle.setFont(headerFont)
+//            val headerRow = sheet.createRow(0)
+//            val columns = arrayOf(
+//                "Price",
+//                "amount",
+//                "ProductName"
+//            )
+//            for (i in columns.indices) {
+//                val cell = headerRow.createCell(i)
+//                cell.setCellValue(columns[i])
+//                cell.cellStyle = headerCellStyle
+//            }
+//
+//            var rowNum = 1
+//            for (payment in payments) {
+//                val row = sheet.createRow(rowNum)
+//                row.createCell(0).setCellValue(payment.price.toDouble())
+//                row.createCell(1).setCellValue(payment.amount.toDouble())
+//                row.createCell(2).setCellValue(payment.productName)
+//                rowNum++
+//            }
+//
+//            for (i in columns.indices) {
+//                sheet.autoSizeColumn(i)
+//            }
+//
+//            val file = File.createTempFile("payments_${createDate}", ".xlsx")
+//            file.deleteOnExit()
+//            val fileOut = FileOutputStream(file)
+//            workbook.write(fileOut)
+//            fileOut.close()
+//            workbook.close()
+//            return file
+//        } catch (e: IOException) {
+//            throw IOException(e)
+//        }
+//    }
 }
